Addressed comments and moved Route class to it's own module
