+++ samcli/local/apigw/local_apigw_service.py
@@ -8,7 +8,7 @@ import json
 import logging
 from datetime import datetime
 from time import time
-from typing import List, Optional
+from typing import Any, Dict, List
 
 from flask import Flask, request
 from werkzeug.datastructures import Headers
@@ -18,9 +18,9 @@ from werkzeug.serving import WSGIRequestHandler
 from samcli.commands.local.lib.exceptions import UnsupportedInlineCodeError
 from samcli.lib.providers.provider import Cors
 from samcli.lib.utils.stream_writer import StreamWriter
-from samcli.local.apigw.authorizers.authorizer import Authorizer
 from samcli.local.apigw.authorizers.lambda_authorizer import LambdaAuthorizer
 from samcli.local.apigw.exceptions import LambdaResponseParseException, PayloadFormatVersionValidateException
+from samcli.local.apigw.route import Route
 from samcli.local.events.api_event import (
     ApiGatewayLambdaEvent,
     ApiGatewayV2LambdaEvent,
@@ -36,85 +36,6 @@ from .path_converter import PathConverter
 from .service_error_responses import ServiceErrorResponses
 
 LOG = logging.getLogger(__name__)
-
-
-class Route:
-    API = "Api"
-    HTTP = "HttpApi"
-    ANY_HTTP_METHODS = ["GET", "DELETE", "PUT", "POST", "HEAD", "OPTIONS", "PATCH"]
-
-    def __init__(
-        self,
-        function_name: Optional[str],
-        path: str,
-        methods: List[str],
-        event_type: str = API,
-        payload_format_version: Optional[str] = None,
-        is_default_route: bool = False,
-        operation_name=None,
-        stack_path: str = "",
-        authorizer_name: Optional[str] = None,
-        authorizer_object: Optional[Authorizer] = None,
-        use_default_authorizer: bool = True,
-    ):
-        """
-        Creates an ApiGatewayRoute
-
-        :param list(str) methods: http method
-        :param function_name: Name of the Lambda function this API is connected to
-        :param str path: Path off the base url
-        :param str event_type: Type of the event. "Api" or "HttpApi"
-        :param str payload_format_version: version of payload format
-        :param bool is_default_route: determines if the default route or not
-        :param string operation_name: Swagger operationId for the route
-        :param str stack_path: path of the stack the route is located
-        :param str authorizer_name: the authorizer this route is using, if any
-        :param Authorizer authorizer_object: the authorizer object this route is using, if any
-        :param bool use_default_authorizer: whether or not to use a default authorizer (if defined)
-        """
-        self.methods = self.normalize_method(methods)
-        self.function_name = function_name
-        self.path = path
-        self.event_type = event_type
-        self.payload_format_version = payload_format_version
-        self.is_default_route = is_default_route
-        self.operation_name = operation_name
-        self.stack_path = stack_path
-        self.authorizer_name = authorizer_name
-        self.authorizer_object = authorizer_object
-        self.use_default_authorizer = use_default_authorizer
-
-    def __eq__(self, other):
-        return (
-            isinstance(other, Route)
-            and sorted(self.methods) == sorted(other.methods)
-            and self.function_name == other.function_name
-            and self.path == other.path
-            and self.operation_name == other.operation_name
-            and self.stack_path == other.stack_path
-            and self.authorizer_name == other.authorizer_name
-            and self.authorizer_object == other.authorizer_object
-            and self.use_default_authorizer == other.use_default_authorizer
-        )
-
-    def __hash__(self):
-        route_hash = hash(f"{self.stack_path}-{self.function_name}-{self.path}")
-        for method in sorted(self.methods):
-            route_hash *= hash(method)
-        return route_hash
-
-    def normalize_method(self, methods):
-        """
-        Normalizes Http Methods. Api Gateway allows a Http Methods of ANY. This is a special verb to denote all
-        supported Http Methods on Api Gateway.
-
-        :param list methods: Http methods
-        :return list: Either the input http_method or one of the _ANY_HTTP_METHODS (normalized Http Methods)
-        """
-        methods = [method.upper() for method in methods]
-        if "ANY" in methods:
-            return self.ANY_HTTP_METHODS
-        return methods
 
 
 class CatchAllPathConverter(BaseConverter):
@@ -284,6 +205,77 @@ class LocalApigwService(BaseLocalService):
         # Something went wrong
         self._app.register_error_handler(500, ServiceErrorResponses.lambda_failure_response)
 
+    def _build_v1_context(self, route: Route) -> Dict[str, Any]:
+        """
+        Helper function to a 1.0 request context
+
+        Parameters
+        ----------
+        route: Route
+            The Route object that was invoked
+
+        Returns
+        -------
+        dict
+            JSON object containing context variables
+        """
+        identity = ContextIdentity(source_ip=request.remote_addr)
+
+        protocol = request.environ.get("SERVER_PROTOCOL", "HTTP/1.1")
+        host = request.host
+
+        operation_name = route.operation_name if route.event_type == Route.API else None
+
+        endpoint = PathConverter.convert_path_to_api_gateway(request.endpoint)
+        method = request.method
+
+        context = RequestContext(
+            resource_path=endpoint,
+            http_method=method,
+            stage=self.api.stage_name,
+            identity=identity,
+            path=endpoint,
+            protocol=protocol,
+            domain_name=host,
+            operation_name=operation_name,
+        )
+
+        return context.to_dict()
+
+    def _build_v2_context(self, route: Route) -> Dict[str, Any]:
+        """
+        Helper function to a 2.0 request context
+
+        Parameters
+        ----------
+        route: Route
+            The Route object that was invoked
+
+        Returns
+        -------
+        dict
+            JSON object containing context variables
+        """
+        endpoint = PathConverter.convert_path_to_api_gateway(request.endpoint)
+        method = request.method
+
+        apigw_endpoint = PathConverter.convert_path_to_api_gateway(endpoint)
+        route_key = self._v2_route_key(method, apigw_endpoint, route.is_default_route)
+
+        request_time_epoch = int(time())
+        request_time = datetime.utcnow().strftime("%d/%b/%Y:%H:%M:%S +0000")
+
+        context_http = ContextHTTP(method=method, path=request.path, source_ip=request.remote_addr)
+        context = RequestContextV2(
+            http=context_http,
+            route_key=route_key,
+            stage=self.api.stage_name,
+            request_time_epoch=request_time_epoch,
+            request_time=request_time,
+        )
+
+        return context.to_dict()
+
     def _valid_identity_sources(self, route: Route) -> bool:
         """
         Validates if the route contains all the valid identity sources defined in the route's Lambda Authorizer
@@ -305,42 +297,11 @@ class LocalApigwService(BaseLocalService):
 
         identity_sources = lambda_auth.identity_sources
 
-        endpoint = PathConverter.convert_path_to_api_gateway(request.endpoint)
-        method = request.method
-
-        if lambda_auth.payload_version == LambdaAuthorizer.PAYLOAD_V1:
-            identity = ContextIdentity(source_ip=request.remote_addr)
-
-            protocol = request.environ.get("SERVER_PROTOCOL", "HTTP/1.1")
-            host = request.host
-
-            operation_name = route.operation_name if route.event_type == Route.API else None
-
-            context = RequestContext(
-                resource_path=endpoint,
-                http_method=method,
-                stage=self.api.stage_name,
-                identity=identity,
-                path=endpoint,
-                protocol=protocol,
-                domain_name=host,
-                operation_name=operation_name,
-            ).to_dict()
-        else:
-            apigw_endpoint = PathConverter.convert_path_to_api_gateway(endpoint)
-            route_key = self._v2_route_key(method, apigw_endpoint, route.is_default_route)
-
-            request_time_epoch = int(time())
-            request_time = datetime.utcnow().strftime("%d/%b/%Y:%H:%M:%S +0000")
-
-            context_http = ContextHTTP(method=method, path=request.path, source_ip=request.remote_addr)
-            context = RequestContextV2(
-                http=context_http,
-                route_key=route_key,
-                stage=self.api.stage_name,
-                request_time_epoch=request_time_epoch,
-                request_time=request_time,
-            ).to_dict()
+        context = (
+            self._build_v1_context(route)
+            if lambda_auth.payload_version == LambdaAuthorizer.PAYLOAD_V1
+            else self._build_v2_context(route)
+        )
 
         kwargs = {
             "headers": request.headers,
