+++ tests/unit/commands/local/lib/test_local_api_service.py
@@ -11,7 +11,7 @@ from samcli.lib.providers.api_collector import ApiCollector
 from samcli.lib.providers.api_provider import ApiProvider
 from samcli.commands.local.lib.exceptions import NoApisDefined
 from samcli.commands.local.lib.local_api_service import LocalApiService
-from samcli.local.apigw.local_apigw_service import Route
+from samcli.local.apigw.route import Route
 
 
 class TestLocalApiService_start(TestCase):
